Enumerate mediaDevices after permission is granted
Some browsers will not show any devices until recording permission is granted,
so if we start recording and the device list is empty, enumerate it again so
the user can select a device next time.

In the demo app, if a device is selected, the recording stops and must be
restarted.